requirements.txt
langchain 
pinecone-client 
langchain-cohere 
langchain-community
streamlit
python-dotenv 
PyPDF2
pypdf